+++ lib.esm/contract/contract.js
@@ -456,6 +456,10 @@ async function _emit(contract, event, args, payloadFunc) {
         catch (error) { }
         return !once;
     });
+    if (sub.listeners.length === 0) {
+        sub.stop();
+        getInternal(contract).subs.delete(sub.tag);
+    }
     return (count > 0);
 }
 async function emit(contract, event, args, payloadFunc) {
@@ -539,7 +543,9 @@ export class BaseContract {
                 }
                 addrPromise = resolver.resolveName(target).then((addr) => {
                     if (addr == null) {
-                        throw new Error("TODO");
+                        throw makeError("an ENS name used for a contract target must be correctly configured", "UNCONFIGURED_NAME", {
+                            value: target
+                        });
                     }
                     getInternal(this).addr = addr;
                     return addr;